CIntRange rounds result when calculating alpha
git-svn-id: http://b035a0a.online-server.cloud/svn/Acf@4810 5144ca80-2444-0410-90ef-8152f95a0219

+++ Include/istd/TRange.h
@@ -637,7 +637,7 @@ inline ValueType TRange<ValueType>::GetValueFromAlpha(double alpha) const
 template <typename ValueType>
 inline double TRange<ValueType>::GetAlphaFromValue(ValueType value) const
 {
-	return (value - GetMinValue()) / GetLength();
+	return static_cast<double>(value - GetMinValue()) / static_cast<double>(GetLength());
 }
